requirements.txt
streamlit==1.31.1
yfinance==0.2.36
numpy==1.26.4
pandas==2.1.4
plotly==5.18.0
tensorflow-cpu==2.15.0
scikit-learn==1.3.2
